Raise LP information out of provider
